Added aging system for creatures
Neat version of the game of life!

I added a class to keep track of the creature's age. Now he grows from a child into an adult through a few steps.

In the future, the creature can have a minimum age to reproduce, an age to die, perhaps even a type etc. to make the game a bit more interesting. 